gpu: Preliminary EGL multi-device support for Linux
This enables the basic EGL multi-device support infrastructure to allow
surfaceless-based Ozone platforms (e.g. Wayland) to use it as a single
source of truth for which DRM device should be used to create the GBM
device with. Since the default device ID also gets passed to ANGLE
during display creation, this effectively ensures that both GBM and
ANGLE is using the same DRM device, avoiding accidental cross-device
resource sharing.