ozone/wayland: Ensure that wl_pointer::enter event updates the cursor
The WaylandWindowManager's current pointer focused window is updated
directly by WaylandDataDragController and WaylandWindowDragController.
This causes the cursor to not be updated on drop as
OnPointerFocusChanged might not be called on wl_pointer::enter event.

Fix this by only updating the cursor when a wl_pointer::event actually
happens instead of when notified by WaylandWindowManager.