ozone/wayland: Defer destruction of clipboard data source on replace
Chromium's current behavior of destroying the current clipboard data
source and immediately replacing it with a new one is causing race
condition issues with certain clipboard managers.

Fix this by deferring the destruction of the clipboard data source to
after the new data source is set as the clipboard data source.