dns: Treat nss-resolve as nss-dns if resolv.conf is systemd-resolved
nss-resolve resolves DNS names through systemd-resolved, and any system
using systemd-resolved may also configure /etc/resolv.conf to also point
to systemd-resolved, in which case the behavior of nss-dns is identical
to nss-resolve. Therefore we can treat nss-resolve as if it's the same
as nss-dns and allow asynchronous DNS resolution to be used.