[Wayland] Support color management protocol
This change adds support for color-management-v1. This enables rendering
to HDR surfaces on supported platforms. I've verified HDR video playback
is working on KDE Plasma 6.4.2. A WaylandWpColorManagerV1 feature flag
is added and enabled by default, which can be used as a remote kill
switch in case we need to disable the feature. Once this lands, the old
zcr_color_manager_v1 can be removed.

Bug: 375959958
Change-Id: I1b1b959de7baf11e0375825fdc11bf0ab8a05915