gfx::DisplayColorSpaces: Unify raster and composite spaces
The color space that is used during raster and composite will affect the
final results via blending behavior. Blending can happen in three places
* Rasterization
* Compositing in viz
* Compositing in the operating system (CoreAnimation)

These three spaces should be the same. To that end, merge the functions
GetRasterColorSpace and GetCompositingColorSpace in
gfx::DisplayColorSpaces.

Ideally, these spaces should be the same regardless of the content that
is being rasterized or composited.

There is one situation where this isn't possible (without significant
compromises). On Windows, in HDR mode, compositing (in viz) can only be
done to sRGB, Rec2020-PQ, or extended-linear-sRGB. Because there is no
suitable 8-bit WCG option, the only way to accomplish WCG content would
be to rasterize to F16. Don't do this yet, because it could be a
performance regression. Instead, accept that blending will change when
WCG content appears on screen. The details of this trade-off are worked
out in DisplayColorSpaces::GetRasterAndCompositeColorSpace.

The pixel tests Pixel_Canvas2DRedBoxScrgbLinear and
Pixel_Canvas2DRedBoxScrgbLinear are affected by this change (they now
blend in sRGB space as opposed to P3).

Bug: 40277679
Change-Id: I602aba292f1deb0d9e3da931564ae366a7281b17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6795085
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1493694}